issue-2559: introduced ShardIdSelectionInLeaderEnabled flag and refactored service_ut to run most of the sharding tests with and without this flag; actual implementation of this flag will be added in the next PR
