+++ cloud/filestore/libs/storage/service/service_ut.cpp
@@ -43,6 +43,41 @@ public:
     }
 };
 
+////////////////////////////////////////////////////////////////////////////////
+
+TString GenerateValidateData(ui32 size)
+{
+    TString data(size, 0);
+    for (ui32 i = 0; i < size; ++i) {
+        data[i] = 'A' + (i % ('Z' - 'A' + 1));
+    }
+    return data;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+NProtoPrivate::TChangeStorageConfigResponse ExecuteChangeStorageConfig(
+    NProto::TStorageConfig config,
+    TServiceClient& service,
+    bool mergeWithConfig = false)
+{
+    NProtoPrivate::TChangeStorageConfigRequest request;
+    request.SetFileSystemId("test");
+
+    *request.MutableStorageConfig() = std::move(config);
+    request.SetMergeWithStorageConfigFromTabletDB(mergeWithConfig);
+
+    TString buf;
+    google::protobuf::util::MessageToJsonString(request, &buf);
+
+    auto jsonResponse = service.ExecuteAction(
+        "changestorageconfig", buf);
+    NProtoPrivate::TChangeStorageConfigResponse response;
+    UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
+        jsonResponse->Record.GetOutput(), &response).ok());
+    return response;
+}
+
 }   // namespace
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -448,7 +483,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
         UNIT_ASSERT_VALUES_EQUAL(profile2.GetBurstPercentage(), 10);
         UNIT_ASSERT_VALUES_EQUAL(profile2.GetMaxWriteCostMultiplier(), 20);
         UNIT_ASSERT_VALUES_EQUAL(profile2.GetDefaultPostponedRequestWeight(), 4_KB);
-
 
         service.AssertDescribeFileStoreModelFailed(0);
         service.AssertDescribeFileStoreModelFailed(1000, 0);
@@ -1355,28 +1389,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
         UNIT_ASSERT_VALUES_EQUAL(
             storageValues.at("CompactionThreshold"),
             "1000");
-    }
-
-    NProtoPrivate::TChangeStorageConfigResponse ExecuteChangeStorageConfig(
-        NProto::TStorageConfig config,
-        TServiceClient& service,
-        bool mergeWithConfig = false)
-    {
-        NProtoPrivate::TChangeStorageConfigRequest request;
-        request.SetFileSystemId("test");
-
-        *request.MutableStorageConfig() = std::move(config);
-        request.SetMergeWithStorageConfigFromTabletDB(mergeWithConfig);
-
-        TString buf;
-        google::protobuf::util::MessageToJsonString(request, &buf);
-
-        auto jsonResponse = service.ExecuteAction(
-            "changestorageconfig", buf);
-        NProtoPrivate::TChangeStorageConfigResponse response;
-        UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-            jsonResponse->Record.GetOutput(), &response).ok());
-        return response;
     }
 
     void CheckStorageConfigValues(
@@ -2320,15 +2332,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
         UNIT_ASSERT_VALUES_EQUAL(4, reassignedChannels[1]);
     }
 
-    TString GenerateValidateData(ui32 size)
-    {
-        TString data(size, 0);
-        for (ui32 i = 0; i < size; ++i) {
-            data[i] = 'A' + (i % ('Z' - 'A' + 1));
-        }
-        return data;
-    }
-
     void CheckThreeStageWrites(NProto::EStorageMediaKind kind, bool disableForHdd)
     {
         TTestEnv env;
@@ -2844,1329 +2847,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
             addData.GetApproximateFreeSpaceShares(0));
     }
 
-    void WaitForTabletStart(TServiceClient& service)
-    {
-        TDispatchOptions options;
-        options.FinalEvents = {
-            TDispatchOptions::TFinalEventCondition(
-                TEvIndexTabletPrivate::EvLoadCompactionMapChunkRequest)};
-        service.AccessRuntime().DispatchEvents(options);
-    }
-
-    void ConfigureShards(
-        TServiceClient& service,
-        const TString& fsId,
-        const TString& shard1Id,
-        const TString& shard2Id)
-    {
-        {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(shard1Id);
-            request.SetShardNo(1);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureasshard", buf);
-            NProtoPrivate::TConfigureAsShardResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(shard2Id);
-            request.SetShardNo(2);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureasshard", buf);
-            NProtoPrivate::TConfigureAsShardResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard1Id;
-            *request.AddShardFileSystemIds() = shard2Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureshards", buf);
-            NProtoPrivate::TConfigureShardsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-    }
-
-    Y_UNIT_TEST(ShouldCreateSessionInShards)
-    {
-        NProto::TStorageConfig config;
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        ui64 nodeId1 =
-            service
-                .CreateNode(headers1, TCreateNodeArgs::File(RootNodeId, "file"))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        ui64 handle1 =
-            service
-                .CreateHandle(
-                    headers1,
-                    headers1.FileSystemId,
-                    nodeId1,
-                    "",
-                    TCreateHandleArgs::RDWR)
-                ->Record.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle1 >= (1LU << 56U) && handle1 < (2LU << 56U),
-            handle1);
-
-        service.WriteData(
-            headers1,
-            headers1.FileSystemId,
-            nodeId1,
-            handle1,
-            0,
-            TString(1_MB, 'a'));
-
-        ui64 nodeId2 =
-            service
-                .CreateNode(headers2, TCreateNodeArgs::File(RootNodeId, "file"))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL((2LU << 56U) + 2, nodeId2);
-
-        ui64 handle2 =
-            service
-                .CreateHandle(
-                    headers2,
-                    headers2.FileSystemId,
-                    nodeId2,
-                    "",
-                    TCreateHandleArgs::RDWR)
-                ->Record.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle2 >= (2LU << 56U) && handle2 < (3LU << 56U),
-            handle2);
-
-        service.WriteData(
-            headers2,
-            headers2.FileSystemId,
-            nodeId2,
-            handle2,
-            0,
-            TString(1_MB, 'a'));
-
-        for (const auto& shardId: {shard1Id, shard2Id}) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(shardId);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(1, sessions.size());
-
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.SessionId,
-                sessions[0].GetSessionId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.ClientId,
-                sessions[0].GetClientId());
-            UNIT_ASSERT_VALUES_EQUAL("", sessions[0].GetSessionState());
-        }
-
-        const TString sessionState = "some_state";
-        service.ResetSession(headers, sessionState);
-
-        for (const auto& shardId: {shard1Id, shard2Id}) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(shardId);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(1, sessions.size());
-
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.SessionId,
-                sessions[0].GetSessionId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.ClientId,
-                sessions[0].GetClientId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                sessionState,
-                sessions[0].GetSessionState());
-        }
-
-        service.DestroySession(headers);
-
-        for (const auto& shardId: {shard1Id, shard2Id}) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(shardId);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(0, sessions.size());
-        }
-    }
-
-    Y_UNIT_TEST(ShouldRestoreSessionInShardAfterShardRestart)
-    {
-        const auto idleSessionTimeout = TDuration::Minutes(2);
-        NProto::TStorageConfig config;
-        config.SetIdleSessionTimeout(idleSessionTimeout.MilliSeconds());
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        TActorId leaderActorId;
-        ui64 shard1TabletId = -1;
-        ui64 shard2TabletId = -1;
-        const auto startupEventType =
-            TEvIndexTabletPrivate::EvLoadCompactionMapChunkRequest;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvDescribeFileStoreResponse: {
-                        using TDesc = TEvSSProxy::TEvDescribeFileStoreResponse;
-                        const auto* msg = event->Get<TDesc>();
-                        const auto& desc =
-                            msg->PathDescription.GetFileStoreDescription();
-                        if (desc.GetConfig().GetFileSystemId() == shard1Id) {
-                            shard1TabletId = desc.GetIndexTabletId();
-                        }
-
-                        if (desc.GetConfig().GetFileSystemId() == shard2Id) {
-                            shard2TabletId = desc.GetIndexTabletId();
-                        }
-
-                        break;
-                    }
-
-                    case startupEventType: {
-                        if (!leaderActorId) {
-                            leaderActorId = event->Sender;
-                        }
-
-                        break;
-                    }
-                }
-
-                return false;
-            });
-
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        // leaderActorId should be re-captured after the restart triggered by
-        // configureshards
-        leaderActorId = {};
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        // creating nodes and handles in both shards
-
-        ui64 nodeId1 =
-            service
-                .CreateNode(headers1, TCreateNodeArgs::File(RootNodeId, "file"))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        ui64 handle1 =
-            service
-                .CreateHandle(
-                    headers1,
-                    headers1.FileSystemId,
-                    nodeId1,
-                    "",
-                    TCreateHandleArgs::RDWR)
-                ->Record.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle1 >= (1LU << 56U) && handle1 < (2LU << 56U),
-            handle1);
-
-        ui64 nodeId2 =
-            service
-                .CreateNode(headers2, TCreateNodeArgs::File(RootNodeId, "file"))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL((2LU << 56U) + 2, nodeId2);
-
-        ui64 handle2 =
-            service
-                .CreateHandle(
-                    headers2,
-                    headers2.FileSystemId,
-                    nodeId2,
-                    "",
-                    TCreateHandleArgs::RDWR)
-                ->Record.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle2 >= (2LU << 56U) && handle2 < (3LU << 56U),
-            handle2);
-
-        // rebooting shards
-
-        TVector<TActorId> fsActorIds;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvIndexTabletPrivate::EvLoadCompactionMapChunkRequest: {
-                        // catching one of the startup events to detect shard
-                        // actor ids
-
-                        if (Find(fsActorIds, event->Sender) == fsActorIds.end()) {
-                            fsActorIds.push_back(event->Sender);
-                        }
-
-                        break;
-                    }
-                }
-
-                return false;
-            });
-
-        TIndexTabletClient shard1(
-            env.GetRuntime(),
-            nodeIdx,
-            shard1TabletId,
-            {}, // config
-            false // updateConfig
-        );
-        shard1.RebootTablet();
-
-        TIndexTabletClient shard2(
-            env.GetRuntime(),
-            nodeIdx,
-            shard2TabletId,
-            {}, // config
-            false // updateConfig
-        );
-        shard2.RebootTablet();
-
-        // triggering shard sessions sync
-        // sending the event manually since registration observers which enable
-        // scheduling for actors are reset upon tablet reboot
-
-        env.GetRuntime().AdvanceCurrentTime(idleSessionTimeout / 2);
-
-        {
-            using TRequest = TEvIndexTabletPrivate::TEvSyncSessionsRequest;
-
-            env.GetRuntime().Send(
-                new IEventHandle(
-                    leaderActorId, // recipient
-                    TActorId(), // sender
-                    new TRequest(),
-                    0, // flags
-                    0),
-                0);
-        }
-
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        // waiting for idle session expiration
-        // sending the event manually since registration observers which enable
-        // scheduling for actors are reset upon tablet reboot
-
-        env.GetRuntime().AdvanceCurrentTime(idleSessionTimeout / 2);
-        service.PingSession(headers);
-
-        for (const auto& actorId: fsActorIds) {
-            using TRequest = TEvIndexTabletPrivate::TEvCleanupSessionsRequest;
-
-            env.GetRuntime().Send(
-                new IEventHandle(
-                    actorId, // recipient
-                    TActorId(), // sender
-                    new TRequest(),
-                    0, // flags
-                    0),
-                0);
-        }
-
-        // need to pass deadline instead of timeout here since otherwise the
-        // adjusted time gets added to the timeout
-        env.GetRuntime().DispatchEvents(
-            {},
-            TInstant::Now() + TDuration::MilliSeconds(100));
-
-        // shard sessions should exist
-
-        for (const auto& id: {fsId, shard1Id, shard2Id}) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(id);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(1, sessions.size());
-
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.SessionId,
-                sessions[0].GetSessionId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.ClientId,
-                sessions[0].GetClientId());
-        }
-
-        // handles should be alive
-
-        service.WriteData(
-            headers1,
-            headers1.FileSystemId,
-            nodeId1,
-            handle1,
-            0,
-            TString(1_MB, 'a'));
-
-        service.WriteData(
-            headers2,
-            headers2.FileSystemId,
-            nodeId2,
-            handle2,
-            0,
-            TString(1_MB, 'a'));
-    }
-
-    Y_UNIT_TEST(ShouldCreateNodeInShardViaLeader)
-    {
-        NProto::TStorageConfig config;
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        ui64 nodeId1 =
-            service
-                .CreateNode(headers, TCreateNodeArgs::File(
-                    RootNodeId,
-                    "file1",
-                    0, // mode
-                    shard1Id))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::RDWR)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            shard1Id,
-            createHandleResponse.GetShardFileSystemId());
-
-        UNIT_ASSERT_VALUES_UNEQUAL(
-            "",
-            createHandleResponse.GetShardNodeName());
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        createHandleResponse = service.CreateHandle(
-            headers1,
-            headers1.FileSystemId,
-            RootNodeId,
-            createHandleResponse.GetShardNodeName(),
-            TCreateHandleArgs::RDWR)->Record;
-
-        auto handle1 = createHandleResponse.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle1 >= (1LU << 56U) && handle1 < (2LU << 56U),
-            handle1);
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse.GetNodeAttr().GetId());
-
-        service.WriteData(
-            headers1,
-            headers1.FileSystemId,
-            nodeId1,
-            handle1,
-            0,
-            TString(1_MB, 'a'));
-    }
-
-    Y_UNIT_TEST(ShouldCreateNodeInShardByCreateHandleViaLeader)
-    {
-        NProto::TStorageConfig config;
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE,
-            shard1Id)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            shard1Id,
-            createHandleResponse.GetShardFileSystemId());
-
-        const auto shardNodeName =
-            createHandleResponse.GetShardNodeName();
-
-        UNIT_ASSERT_VALUES_UNEQUAL("", shardNodeName);
-
-        const auto nodeId1 = createHandleResponse.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        createHandleResponse = service.CreateHandle(
-            headers1,
-            headers1.FileSystemId,
-            RootNodeId,
-            shardNodeName,
-            TCreateHandleArgs::RDWR)->Record;
-
-        auto handle1 = createHandleResponse.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle1 >= (1LU << 56U) && handle1 < (2LU << 56U),
-            handle1);
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse.GetNodeAttr().GetId());
-
-        service.WriteData(
-            headers1,
-            headers1.FileSystemId,
-            nodeId1,
-            handle1,
-            0,
-            TString(1_MB, 'a'));
-
-        auto getAttrResponse = service.GetNodeAttr(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1")->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            shard1Id,
-            getAttrResponse.GetNode().GetShardFileSystemId());
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            shardNodeName,
-            getAttrResponse.GetNode().GetShardNodeName());
-
-        getAttrResponse = service.GetNodeAttr(
-            headers1,
-            shard1Id,
-            RootNodeId,
-            shardNodeName)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(nodeId1, getAttrResponse.GetNode().GetId());
-        UNIT_ASSERT_VALUES_EQUAL(1_MB, getAttrResponse.GetNode().GetSize());
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            shard1Id,
-            listNodesResponse.GetNodes(0).GetShardFileSystemId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            shardNodeName,
-            listNodesResponse.GetNodes(0).GetShardNodeName());
-    }
-
-    Y_UNIT_TEST(ShouldForwardRequestsToShard)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE_EXL)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            "",
-            createHandleResponse.GetShardFileSystemId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            "",
-            createHandleResponse.GetShardNodeName());
-
-        const auto nodeId1 = createHandleResponse.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        const auto handle1 = createHandleResponse.GetHandle();
-
-        UNIT_ASSERT_C(
-            handle1 >= (1LU << 56U) && handle1 < (2LU << 56U),
-            handle1);
-
-        auto accessNodeResponse = service.AccessNode(
-            headers,
-            fsId,
-            nodeId1)->Record;
-
-        Y_UNUSED(accessNodeResponse);
-
-        auto setNodeAttrResponse = service.SetNodeAttr(
-            headers,
-            fsId,
-            nodeId1,
-            1_MB)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            setNodeAttrResponse.GetNode().GetId());
-
-        UNIT_ASSERT_VALUES_EQUAL(1_MB, setNodeAttrResponse.GetNode().GetSize());
-
-        auto allocateDataResponse = service.AllocateData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1,
-            0,
-            2_MB)->Record;
-
-        Y_UNUSED(allocateDataResponse);
-
-        auto data = GenerateValidateData(256_KB);
-        service.WriteData(headers, fsId, nodeId1, handle1, 0, data);
-        auto readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1,
-            0,
-            data.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data, readDataResponse.GetBuffer());
-
-        auto destroyHandleResponse = service.DestroyHandle(
-            headers,
-            fsId,
-            nodeId1,
-            handle1)->Record;
-
-        Y_UNUSED(destroyHandleResponse);
-
-        const auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file2"))->Record;
-
-        const auto nodeId2 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((2LU << 56U) + 2, nodeId2);
-
-        auto getNodeAttrResponse = service.GetNodeAttr(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1")->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            getNodeAttrResponse.GetNode().GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            2_MB,
-            getNodeAttrResponse.GetNode().GetSize());
-
-        getNodeAttrResponse = service.GetNodeAttr(
-            headers,
-            fsId,
-            nodeId1,
-            "")->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            getNodeAttrResponse.GetNode().GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            2_MB,
-            getNodeAttrResponse.GetNode().GetSize());
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(1));
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            2_MB,
-            listNodesResponse.GetNodes(0).GetSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId2,
-            listNodesResponse.GetNodes(1).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            0,
-            listNodesResponse.GetNodes(1).GetSize());
-
-        service.UnlinkNode(headers, RootNodeId, "file1");
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(0));
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId2,
-            listNodesResponse.GetNodes(0).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            0,
-            listNodesResponse.GetNodes(0).GetSize());
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        listNodesResponse = service.ListNodes(
-            headers2,
-            shard2Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-
-        auto setXAttrResponse = service.SetNodeXAttr(
-            headers,
-            fsId,
-            nodeId2,
-            "user.some_attr",
-            "some_value")->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, setXAttrResponse.GetVersion());
-
-        auto getXAttrResponse = service.GetNodeXAttr(
-            headers,
-            fsId,
-            nodeId2,
-            "user.some_attr")->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, getXAttrResponse.GetVersion());
-        UNIT_ASSERT_VALUES_EQUAL("some_value", getXAttrResponse.GetValue());
-
-        service.SendSetNodeXAttrRequest(
-            headers,
-            fsId,
-            nodeId1,
-            "user.some_attr",
-            "some_value");
-
-        auto setNodeXAttrResponseEvent = service.RecvSetNodeXAttrResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_FS_NOENT,
-            setNodeXAttrResponseEvent->GetStatus(),
-            setNodeXAttrResponseEvent->GetErrorReason());
-    }
-
-    Y_UNIT_TEST(ShouldCreateDirectoryStructureInLeader)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::Directory(RootNodeId, "dir1"))->Record;
-        const auto dir1Id = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(0, ExtractShardNo(dir1Id));
-
-        createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::Directory(dir1Id, "dir1_1"))->Record;
-        const auto dir1_1Id = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(0, ExtractShardNo(dir1_1Id));
-
-        createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::Directory(dir1Id, "dir1_2"))->Record;
-        const auto dir1_2Id = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(0, ExtractShardNo(dir1_2Id));
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            dir1_1Id,
-            "file1",
-            TCreateHandleArgs::CREATE)->Record;
-
-        const auto nodeId1 = createHandleResponse.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(ShardedId(2, 1), nodeId1);
-
-        const auto handle1 = createHandleResponse.GetHandle();
-        UNIT_ASSERT_C(ExtractShardNo(handle1) == 1, handle1);
-
-        createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            dir1_2Id,
-            "file1",
-            TCreateHandleArgs::CREATE)->Record;
-
-        const auto nodeId2 = createHandleResponse.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(ShardedId(2, 2), nodeId2);
-
-        const auto handle2 = createHandleResponse.GetHandle();
-        UNIT_ASSERT_C(ExtractShardNo(handle2) == 2, handle2);
-
-        auto data = GenerateValidateData(256_KB);
-        service.WriteData(headers, fsId, nodeId1, handle1, 0, data);
-        auto readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1,
-            0,
-            data.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data, readDataResponse.GetBuffer());
-
-        data = GenerateValidateData(1_MB);
-        service.WriteData(headers, fsId, nodeId2, handle2, 0, data);
-        readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId2,
-            handle2,
-            0,
-            data.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data, readDataResponse.GetBuffer());
-
-        service.DestroyHandle(headers, fsId, nodeId1, handle1);
-        service.DestroyHandle(headers, fsId, nodeId2, handle2);
-    }
-
-    Y_UNIT_TEST(ShouldNotFailOnLegacyHandlesWithHighBits)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-
-        // forcing fs to set high bits for handle ids
-        {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(fsId);
-            request.SetShardNo(111);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureasshard", buf);
-            NProtoPrivate::TConfigureAsShardResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE_EXL)->Record;
-
-        const auto nodeId1 = createHandleResponse.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(111, ExtractShardNo(nodeId1));
-
-        const auto handle1 = createHandleResponse.GetHandle();
-        UNIT_ASSERT_VALUES_EQUAL(111, ExtractShardNo(handle1));
-
-        auto data = GenerateValidateData(256_KB);
-        service.WriteData(headers, fsId, nodeId1, handle1, 0, data);
-        auto readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1,
-            0,
-            data.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data, readDataResponse.GetBuffer());
-    }
-
-    Y_UNIT_TEST(ShouldHandleCreateNodeErrorFromShardUponCreateHandleViaLeader)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto error = MakeError(E_FAIL);
-
-        env.GetRuntime().SetEventFilter(
-            [&] (TTestActorRuntimeBase&, TAutoPtr<IEventHandle>& event) {
-                switch (event->GetTypeRewrite()) {
-                    case TEvService::EvCreateNodeResponse: {
-                        auto* msg =
-                            event->Get<TEvService::TEvCreateNodeResponse>();
-                        if (error.GetCode()) {
-                            msg->Record.MutableError()->CopyFrom(error);
-                        }
-
-                        break;
-                    }
-                }
-
-                return false;
-            });
-
-        const ui64 requestId = 111;
-
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE_EXL,
-            "",
-            requestId);
-
-        auto response = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            error.GetCode(),
-            response->GetError().GetCode(),
-            FormatError(response->GetError()));
-
-        error = {};
-
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE_EXL,
-            "",
-            requestId);
-
-        response = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            error.GetCode(),
-            response->GetError().GetCode(),
-            FormatError(response->GetError()));
-    }
-
-    void DoTestShouldNotFailListNodesUponGetAttrENOENT(
-        NProto::TStorageConfig config)
-    {
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"));
-        service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file2"));
-        service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file3"));
-        service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file4"));
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(4, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(1));
-        UNIT_ASSERT_VALUES_EQUAL("file3", listNodesResponse.GetNames(2));
-        UNIT_ASSERT_VALUES_EQUAL("file4", listNodesResponse.GetNames(3));
-        TVector<std::pair<ui64, TString>> nodes(4);
-        for (ui32 i = 0; i < 4; ++i) {
-            nodes[i] = {
-                listNodesResponse.GetNodes(i).GetId(),
-                listNodesResponse.GetNames(i)};
-            UNIT_ASSERT_VALUES_UNEQUAL(0, nodes[i].first);
-        }
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NamesSize());
-        const auto shard1NodeName1 = listNodesResponse.GetNames(0);
-        const auto shard1NodeId1 = listNodesResponse.GetNodes(0).GetId();
-        const auto shard1NodeName2 = listNodesResponse.GetNames(1);
-
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        listNodesResponse = service.ListNodes(
-            headers2,
-            shard2Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NamesSize());
-        const auto shard2NodeName1 = listNodesResponse.GetNames(0);
-        const auto shard2NodeName2 = listNodesResponse.GetNames(1);
-
-        // "breaking" one node - deleting it directly from the shard
-        service.UnlinkNode(headers1, RootNodeId, shard1NodeName1);
-
-        EraseIf(nodes, [=] (const auto& node) {
-            return node.first == shard1NodeId1;
-        });
-
-        // ListNodes should still succeed
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        // unresolved nodes should be removed from the response
-        UNIT_ASSERT_VALUES_EQUAL(3, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[0].second,
-            listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[1].second,
-            listNodesResponse.GetNames(1));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[2].second,
-            listNodesResponse.GetNames(2));
-
-        UNIT_ASSERT_VALUES_EQUAL(3, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[0].first,
-            listNodesResponse.GetNodes(0).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[1].first,
-            listNodesResponse.GetNodes(1).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodes[2].first,
-            listNodesResponse.GetNodes(2).GetId());
-
-        const auto counters =
-            env.GetCounters()->FindSubgroup("component", "service");
-        UNIT_ASSERT(counters);
-        const auto counter =
-            counters->GetCounter("AppCriticalEvents/NodeNotFoundInShard");
-        UNIT_ASSERT_EQUAL(1, counter->GetAtomic());
-
-        // "breaking" all nodes - ListNodes should fail with E_IO after this
-        service.UnlinkNode(headers1, RootNodeId, shard1NodeName2);
-        service.UnlinkNode(headers2, RootNodeId, shard2NodeName1);
-        service.UnlinkNode(headers2, RootNodeId, shard2NodeName2);
-
-        service.SendListNodesRequest(headers, fsId, RootNodeId);
-        auto response = service.RecvListNodesResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_IO,
-            response->GetError().GetCode(),
-            response->GetErrorReason());
-    }
-
-    Y_UNIT_TEST(ShouldNotFailListNodesUponGetAttrENOENT)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        DoTestShouldNotFailListNodesUponGetAttrENOENT(std::move(config));
-    }
-
-    Y_UNIT_TEST(ShouldNotFailListNodesUponGetAttrENOENTWithGetNodeAttrBatch)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        config.SetGetNodeAttrBatchEnabled(true);
-        DoTestShouldNotFailListNodesUponGetAttrENOENT(std::move(config));
-    }
-
-    Y_UNIT_TEST(ShouldListMultipleNodesWithGetNodeAttrBatch)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        config.SetGetNodeAttrBatchEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        TVector<TString> names;
-        TVector<ui64> ids;
-        for (ui32 i = 10; i < 50; ++i) {
-            const auto name = Sprintf("file%u", i);
-            const auto id = service.CreateNode(
-                headers,
-                TCreateNodeArgs::File(RootNodeId, name)
-            )->Record.GetNode().GetId();
-            names.push_back(name);
-            ids.push_back(id);
-        }
-
-        auto listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            names.size() / 2,
-            listNodesResponse.NamesSize());
-        const ui32 idxToUnlink1 = 13;
-        const ui32 idxToUnlink2 = 17;
-        const auto shard1NodeName1 = listNodesResponse.GetNames(idxToUnlink1);
-        const ui64 unlinkedId1 = listNodesResponse.GetNodes(idxToUnlink1).GetId();
-        const auto shard1NodeName2 = listNodesResponse.GetNames(idxToUnlink2);
-        const ui64 unlinkedId2 = listNodesResponse.GetNodes(idxToUnlink2).GetId();
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(names.size(), listNodesResponse.NamesSize());
-        for (ui32 i = 0; i < names.size(); ++i) {
-            UNIT_ASSERT_VALUES_EQUAL(names[i], listNodesResponse.GetNames(i));
-            UNIT_ASSERT_VALUES_EQUAL(
-                ids[i],
-                listNodesResponse.GetNodes(i).GetId());
-        }
-
-        service.UnlinkNode(headers1, RootNodeId, shard1NodeName1);
-        service.UnlinkNode(headers1, RootNodeId, shard1NodeName2);
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        for (auto id: {unlinkedId1, unlinkedId2}) {
-            auto idx = Find(ids, id) - ids.begin();
-            ids.erase(ids.begin() + idx);
-            names.erase(names.begin() + idx);
-        }
-
-        UNIT_ASSERT_VALUES_EQUAL(names.size(), listNodesResponse.NamesSize());
-        for (ui32 i = 0; i < names.size(); ++i) {
-            UNIT_ASSERT_VALUES_EQUAL(names[i], listNodesResponse.GetNames(i));
-            UNIT_ASSERT_VALUES_EQUAL(
-                ids[i],
-                listNodesResponse.GetNodes(i).GetId());
-        }
-    }
-
     Y_UNIT_TEST(DestroyFileStoreWithActiveSessionShouldFail)
     {
         TTestEnv env;
@@ -4275,604 +2955,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
         UNIT_ASSERT_VALUES_EQUAL(fsId, fsIds[0]);
     }
 
-    Y_UNIT_TEST(ShouldValidateRequestsWithShardId)
-    {
-        TTestEnv env;
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        const auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse.GetNode().GetId();
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        // a request with ShardId and without Name
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            nodeId1,
-            "",
-            TCreateHandleArgs::RDWR,
-            shard1Id);
-
-        auto createHandleResponseEvent = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_ARGUMENT,
-            createHandleResponseEvent->GetError().GetCode(),
-            createHandleResponseEvent->GetErrorReason());
-    }
-
-    Y_UNIT_TEST(ShouldValidateShardConfiguration)
-    {
-        TTestEnv env;
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-        const auto shard3Id = fsId + "-f3";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-        service.CreateFileStore(shard3Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        // ShardNo change not allowed
-        {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(shard1Id);
-            request.SetShardNo(2);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureasshard", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_ARGUMENT,
-                response->GetError().GetCode(),
-                response->GetErrorReason());
-        }
-
-        // Shard deletion not allowed
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard1Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_ARGUMENT,
-                response->GetError().GetCode(),
-                response->GetErrorReason());
-        }
-
-        // Shard reordering not allowed
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard2Id;
-            *request.AddShardFileSystemIds() = shard1Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_ARGUMENT,
-                response->GetError().GetCode(),
-                response->GetErrorReason());
-        }
-
-        // Shard addition IS allowed
-        {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(shard3Id);
-            request.SetShardNo(3);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureasshard", buf);
-            NProtoPrivate::TConfigureAsShardResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard1Id;
-            *request.AddShardFileSystemIds() = shard2Id;
-            *request.AddShardFileSystemIds() = shard3Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureshards", buf);
-            NProtoPrivate::TConfigureShardsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        // TODO(#1350): leader should check that shards' ShardNos correspond
-        // to the shard order in leader's config
-    }
-
-    Y_UNIT_TEST(ShouldRenameExternalNodes)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        // creating 2 files
-
-        auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(1, ExtractShardNo(nodeId1));
-
-        createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file2"))->Record;
-
-        const auto nodeId2 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(2, ExtractShardNo(nodeId2));
-
-        ui64 handle1 = service.CreateHandle(
-            headers,
-            fsId,
-            nodeId1,
-            "",
-            TCreateHandleArgs::RDWR)->Record.GetHandle();
-
-        // writing some data to file1 then moving file1 to file3
-
-        auto data1 = GenerateValidateData(256_KB);
-        service.WriteData(headers, fsId, nodeId1, handle1, 0, data1);
-
-        auto renameNodeResponse = service.RenameNode(
-            headers,
-            RootNodeId,
-            "file1",
-            RootNodeId,
-            "file3",
-            0);
-
-        // opening file3 for reading
-
-        auto createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file3",
-            TCreateHandleArgs::RDNLY)->Record;
-
-        // checking that file3 refers to the same node as formerly file1
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse.GetNodeAttr().GetId());
-
-        auto handle1r = createHandleResponse.GetHandle();
-
-        // checking that we can read the data that we wrote to file1
-
-        auto readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1r,
-            0,
-            data1.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data1, readDataResponse.GetBuffer());
-
-        // checking that node listing shows file3 and file2
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL("file3", listNodesResponse.GetNames(1));
-
-        UNIT_ASSERT_VALUES_EQUAL(2, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId2,
-            listNodesResponse.GetNodes(0).GetId());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(1).GetId());
-
-        // checking that move into an existing file (file2) works
-
-        renameNodeResponse = service.RenameNode(
-            headers,
-            RootNodeId,
-            "file3",
-            RootNodeId,
-            "file2",
-            0);
-
-        // checking that we can still read the same data
-
-        createHandleResponse = service.CreateHandle(
-            headers,
-            fsId,
-            RootNodeId,
-            "file2",
-            TCreateHandleArgs::RDNLY)->Record;
-
-        // nodeId should be kept intact
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse.GetNodeAttr().GetId());
-
-        handle1r = createHandleResponse.GetHandle();
-
-        readDataResponse = service.ReadData(
-            headers,
-            fsId,
-            nodeId1,
-            handle1r,
-            0,
-            data1.size())->Record;
-        UNIT_ASSERT_VALUES_EQUAL(data1, readDataResponse.GetBuffer());
-
-        // listing should show only file2 now
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(0));
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        // listing in shard2 should show nothing
-
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        listNodesResponse = service.ListNodes(
-            headers2,
-            shard2Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        // now try to move to another subdirectory
-
-        auto subdirId =
-            service
-                .CreateNode(
-                    headers,
-                    TCreateNodeArgs::Directory(RootNodeId, "subdir"))
-                ->Record.GetNode()
-                .GetId();
-
-        renameNodeResponse = service.RenameNode(
-            headers,
-            RootNodeId,
-            "file2",
-            subdirId,
-            "file2",
-            0);
-
-        // listing should show only subdir with file2 in it
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("subdir", listNodesResponse.GetNames(0));
-
-        listNodesResponse = service.ListNodes(headers, fsId, subdirId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(0));
-
-        // create 2 files in the same shard
-
-        ui64 nodeId4 =
-            service
-                .CreateNode(headers, TCreateNodeArgs::File(RootNodeId, "file4"))
-                ->Record.GetNode()
-                .GetId();
-
-        // round robin is used for 2 clusters, so we just skip one shard by
-        // creating a file in it
-        service.CreateNode(headers, TCreateNodeArgs::File(RootNodeId, "file5"));
-
-        ui64 nodeId6 =
-            service
-                .CreateNode(headers, TCreateNodeArgs::File(RootNodeId, "file6"))
-                ->Record.GetNode()
-                .GetId();
-
-        UNIT_ASSERT_VALUES_EQUAL(
-            ExtractShardNo(nodeId4),
-            ExtractShardNo(nodeId6));
-
-        // now move to the same shard
-
-        renameNodeResponse = service.RenameNode(
-            headers,
-            RootNodeId,
-            "file4",
-            RootNodeId,
-            "file6",
-            0);
-
-        // file4 should not be present in the listing
-
-        service.SendGetNodeAttrRequest(headers, fsId, RootNodeId, "file4");
-
-        auto getNodeAttrResponse = service.RecvGetNodeAttrResponse();
-        UNIT_ASSERT(getNodeAttrResponse);
-        UNIT_ASSERT_C(
-            FAILED(getNodeAttrResponse->GetStatus()),
-            getNodeAttrResponse->GetErrorReason().c_str());
-    }
-
-    Y_UNIT_TEST(ShouldPerformLocksForExternalNodes)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        ui64 nodeId = service
-            .CreateNode(headers, TCreateNodeArgs::File(RootNodeId, "file"))
-            ->Record.GetNode()
-            .GetId();
-        ui64 handle = service
-            .CreateHandle(headers, fsId, nodeId, "", TCreateHandleArgs::RDWR)
-            ->Record.GetHandle();
-
-        service.AcquireLock(headers, fsId, handle, 1, 0, 4_KB);
-
-        auto response =
-            service.AssertAcquireLockFailed(headers, fsId, handle, 2, 0, 4_KB);
-        UNIT_ASSERT_VALUES_EQUAL(response->GetError().GetCode(), E_FS_WOULDBLOCK);
-
-        service.ReleaseLock(headers, fsId, handle, 2, 0, 4_KB);
-        service.AcquireLock(headers, fsId, handle, 1, 0, 0);
-
-        response =
-            service.AssertAcquireLockFailed(headers, fsId, handle, 2, 0, 4_KB);
-        UNIT_ASSERT_VALUES_EQUAL(response->GetError().GetCode(), E_FS_WOULDBLOCK);
-
-        service.ReleaseLock(headers, fsId, handle, 1, 0, 0);
-
-        // ShouldTrackSharedLocks
-
-        service.AcquireLock(
-            headers,
-            fsId,
-            handle,
-            1,
-            0,
-            4_KB,
-            DefaultPid,
-            NProto::E_SHARED);
-        service.AcquireLock(
-            headers,
-            fsId,
-            handle,
-            2,
-            0,
-            4_KB,
-            DefaultPid,
-            NProto::E_SHARED);
-
-        response =
-            service.AssertAcquireLockFailed(headers, fsId, handle, 1, 0, 0);
-        UNIT_ASSERT_VALUES_EQUAL(
-            response->GetError().GetCode(),
-            E_FS_WOULDBLOCK);
-
-        response = service.AssertAcquireLockFailed(headers, fsId, handle, 1, 0, 0);
-        UNIT_ASSERT_VALUES_EQUAL(response->GetError().GetCode(), E_FS_WOULDBLOCK);
-
-        service.AcquireLock(headers, fsId, handle, 3, 0, 4_KB, DefaultPid, NProto::E_SHARED);
-
-        service.DestroyHandle(headers, fsId, nodeId, handle);
-
-        handle = service
-            .CreateHandle(headers, fsId, nodeId, "", TCreateHandleArgs::WRNLY)
-            ->Record.GetHandle();
-
-        service.AssertTestLockFailed(headers, fsId, handle, 1, 0, 4_KB, DefaultPid, NProto::E_SHARED);
-        service.AssertAcquireLockFailed(headers, fsId, handle, 1, 0, 4_KB, DefaultPid, NProto::E_SHARED);
-    }
-
-    Y_UNIT_TEST(ShouldLinkExternalNodes)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        const auto nodeId1 =
-            service
-                .CreateNode(headers, TCreateNodeArgs::File(RootNodeId, "file1"))
-                ->Record.GetNode()
-                .GetId();
-        UNIT_ASSERT_VALUES_EQUAL(1, ExtractShardNo(nodeId1));
-
-        auto node = service.GetNodeAttr(headers, fsId, RootNodeId, "file1")
-                        ->Record.GetNode();
-        UNIT_ASSERT_VALUES_EQUAL(1, node.GetLinks());
-
-        auto linkNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::Link(RootNodeId, "file2", nodeId1));
-        UNIT_ASSERT(linkNodeResponse);
-        UNIT_ASSERT_C(
-            SUCCEEDED(linkNodeResponse->GetStatus()),
-            linkNodeResponse->GetErrorReason().c_str());
-
-        // attributes of the linked node should be updated
-        UNIT_ASSERT_VALUES_EQUAL(
-            2,
-            linkNodeResponse->Record.GetNode().GetLinks());
-        UNIT_ASSERT_VALUES_UNEQUAL(
-            0,
-            linkNodeResponse->Record.GetNode().GetCTime());
-
-        // validate that the links field is incremented
-        auto links = service.GetNodeAttr(headers, fsId, RootNodeId, "file1")
-                         ->Record.GetNode().GetLinks();
-        UNIT_ASSERT_VALUES_EQUAL(2, links);
-
-        // get node attr should also work with the link
-        links = service.GetNodeAttr(headers, fsId, RootNodeId, "file1")
-                    ->Record.GetNode().GetLinks();
-        UNIT_ASSERT_VALUES_EQUAL(2, links);
-
-        // validate that reading from hardlinked file works
-        auto data = GenerateValidateData(256_KB);
-        ui64 handle = service
-            .CreateHandle(headers, fsId, RootNodeId, "file1", TCreateHandleArgs::RDWR)
-            ->Record.GetHandle();
-        service.WriteData(headers, fsId, nodeId1, handle, 0, data);
-
-        ui64 handle2 = service
-            .CreateHandle(headers, fsId, RootNodeId, "file2", TCreateHandleArgs::RDWR)
-            ->Record.GetHandle();
-
-        auto data2 = service.ReadData(
-            headers,
-            fsId,
-            linkNodeResponse->Record.GetNode().GetId(),
-            handle2,
-            0,
-            data.size())->Record.GetBuffer();
-        UNIT_ASSERT_VALUES_EQUAL(data, data2);
-
-        // Removal of both the file and a hardlink should remove file from both
-        // the shard and a leader
-
-        service.DestroyHandle(headers, fsId, nodeId1, handle);
-        service.DestroyHandle(headers, fsId, linkNodeResponse->Record.GetNode().GetId(), handle2);
-
-        service.UnlinkNode(headers, RootNodeId, "file1");
-        service.UnlinkNode(headers, RootNodeId, "file2");
-
-        // Now listing of the root should show no files
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        // GetNodeAttr should fail as well
-
-        service.AssertGetNodeAttrFailed(headers, fsId, RootNodeId, "file1");
-        service.AssertGetNodeAttrFailed(headers, fsId, RootNodeId, "file2");
-
-        // Creating hardlinks to non-existing files should fail. It is
-        // reasonable to assume that nodeId + 100 is not a valid node id.
-        linkNodeResponse = service.AssertCreateNodeFailed(
-            headers,
-            TCreateNodeArgs::Link(RootNodeId, "file3", nodeId1 + 100));
-    }
-
     Y_UNIT_TEST(ShouldForceDestroyWithAllowFileStoreForceDestroyFlag)
     {
         NProto::TStorageConfig storageConfig;
@@ -4915,979 +2997,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
             createSessionResponse->GetStatus(),
             createSessionResponse->GetErrorReason());
         service.AssertDestroyFileStoreFailed(fsId, true);
-    }
-
-    Y_UNIT_TEST(ShouldAggregateFileSystemMetrics)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        // creating 2 files
-
-        auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(1, ExtractShardNo(nodeId1));
-
-        createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file2"))->Record;
-
-        const auto nodeId2 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL(2, ExtractShardNo(nodeId2));
-
-        ui64 handle1 = service.CreateHandle(
-            headers,
-            fsId,
-            nodeId1,
-            "",
-            TCreateHandleArgs::RDWR)->Record.GetHandle();
-
-        auto data1 = GenerateValidateData(256_KB);
-        service.WriteData(headers, fsId, nodeId1, handle1, 0, data1);
-
-        ui64 handle2 = service.CreateHandle(
-            headers,
-            fsId,
-            nodeId2,
-            "",
-            TCreateHandleArgs::RDWR)->Record.GetHandle();
-
-        auto data2 = GenerateValidateData(512_KB);
-        service.WriteData(headers, fsId, nodeId2, handle2, 0, data2);
-
-        // waiting for async stats aggregation from shards
-        // doing it before triggering another event to avoid DispatchEvents call
-        // which does a long busy-wait loop
-        env.GetRuntime().AdvanceCurrentTime(TDuration::Seconds(15));
-
-        const auto fsStat = service.StatFileStore(headers, fsId)->Record;
-        const auto& fileStore = fsStat.GetFileStore();
-        UNIT_ASSERT_VALUES_EQUAL(fsId, fileStore.GetFileSystemId());
-        UNIT_ASSERT_VALUES_EQUAL(1'000, fileStore.GetBlocksCount());
-        UNIT_ASSERT_VALUES_EQUAL(4_KB, fileStore.GetBlockSize());
-
-        const auto& fileStoreStats = fsStat.GetStats();
-        UNIT_ASSERT_VALUES_EQUAL(2, fileStoreStats.GetUsedNodesCount());
-        UNIT_ASSERT_VALUES_EQUAL(
-            768_KB / 4_KB,
-            fileStoreStats.GetUsedBlocksCount());
-
-        {
-            NProtoPrivate::TGetStorageStatsRequest request;
-            request.SetFileSystemId(fsId);
-            request.SetAllowCache(true);
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            const auto response = service.ExecuteAction("GetStorageStats", buf);
-            NProtoPrivate::TGetStorageStatsResponse record;
-            auto status = google::protobuf::util::JsonStringToMessage(
-                response->Record.GetOutput(),
-                &record);
-            const auto& stats = record.GetStats();
-            UNIT_ASSERT_VALUES_EQUAL(
-                (data1.size() + data2.size()) / 4_KB,
-                stats.GetUsedBlocksCount());
-            UNIT_ASSERT_VALUES_EQUAL(
-                (data1.size() + data2.size()) / 4_KB,
-                stats.GetMixedBlocksCount());
-        }
-
-        service.WriteData(headers, fsId, nodeId2, handle2, 0, data2);
-
-        // waiting for async stats aggregation from shards
-        // doing it before triggering another event to avoid DispatchEvents call
-        // which does a long busy-wait loop
-        env.GetRuntime().AdvanceCurrentTime(TDuration::Seconds(15));
-        // just triggering another event chain - doesn't matter which one
-        service.StatFileStore(headers, fsId);
-
-        {
-            NProtoPrivate::TGetStorageStatsRequest request;
-            request.SetFileSystemId(fsId);
-            request.SetAllowCache(true);
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            const auto response = service.ExecuteAction("GetStorageStats", buf);
-            NProtoPrivate::TGetStorageStatsResponse record;
-            auto status = google::protobuf::util::JsonStringToMessage(
-                response->Record.GetOutput(),
-                &record);
-            const auto& stats = record.GetStats();
-            UNIT_ASSERT_VALUES_EQUAL(
-                (data1.size() + data2.size()) / 4_KB,
-                stats.GetUsedBlocksCount());
-            UNIT_ASSERT_VALUES_EQUAL(
-                (data1.size() + 2 * data2.size()) / 4_KB,
-                stats.GetMixedBlocksCount());
-        }
-    }
-
-    Y_UNIT_TEST(ShouldRetryUnlinkingInShard)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        const auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        TAutoPtr<IEventHandle> shardUnlinkResponse;
-        bool intercept = true;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvUnlinkNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvUnlinkNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        auto response = std::make_unique<
-                            TEvService::TEvUnlinkNodeResponse>(
-                            MakeError(E_REJECTED, "error"));
-
-                        shardUnlinkResponse = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        const ui64 requestId = 111;
-        service.SendUnlinkNodeRequest(
-            headers,
-            RootNodeId,
-            "file1",
-            false, // unlinkDirectory
-            requestId);
-
-        ui32 iterations = 0;
-        while (!shardUnlinkResponse && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(shardUnlinkResponse);
-        intercept = false;
-        env.GetRuntime().Send(shardUnlinkResponse.Release());
-
-        auto unlinkResponse = service.RecvUnlinkNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            unlinkResponse->GetError().GetCode(),
-            unlinkResponse->GetError().GetMessage());
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        // checking DupCache logic - just in case
-        service.SendUnlinkNodeRequest(
-            headers,
-            RootNodeId,
-            "file1",
-            false, // unlinkDirectory
-            requestId);
-
-        unlinkResponse = service.RecvUnlinkNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            unlinkResponse->GetError().GetCode(),
-            unlinkResponse->GetError().GetMessage());
-    }
-
-    Y_UNIT_TEST(ShouldRetryUnlinkingInShardUponLeaderRestart)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        ui64 tabletId = -1;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvDescribeFileStoreResponse: {
-                        using TDesc = TEvSSProxy::TEvDescribeFileStoreResponse;
-                        const auto* msg = event->Get<TDesc>();
-                        const auto& desc =
-                            msg->PathDescription.GetFileStoreDescription();
-                        if (desc.GetConfig().GetFileSystemId() == fsId) {
-                            tabletId = desc.GetIndexTabletId();
-                        }
-                    }
-                }
-
-                return false;
-            });
-
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        const auto createNodeResponse = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        bool intercept = true;
-        bool intercepted = false;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvUnlinkNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvUnlinkNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        intercepted = true;
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        service.SendUnlinkNodeRequest(headers, RootNodeId, "file1");
-
-        ui32 iterations = 0;
-        while (!intercepted && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(intercepted);
-        intercept = false;
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        auto listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-
-        TIndexTabletClient tablet(env.GetRuntime(), nodeIdx, tabletId);
-        tablet.RebootTablet();
-
-        auto unlinkResponse = service.RecvUnlinkNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_REJECTED,
-            unlinkResponse->GetError().GetCode(),
-            unlinkResponse->GetError().GetMessage());
-
-        // remaking session since CreateSessionActor doesn't do it by itself
-        // because EvWakeup never arrives because Scheduling doesn't work by
-        // default and RegistrationObservers get reset after RebootTablet
-        headers = service.InitSession(fsId, "client");
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-    }
-
-    Y_UNIT_TEST(ShouldRetryUnlinkingInShardUponLeaderRestartForRenameNode)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        ui64 tabletId = -1;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvDescribeFileStoreResponse: {
-                        using TDesc = TEvSSProxy::TEvDescribeFileStoreResponse;
-                        const auto* msg = event->Get<TDesc>();
-                        const auto& desc =
-                            msg->PathDescription.GetFileStoreDescription();
-                        if (desc.GetConfig().GetFileSystemId() == fsId) {
-                            tabletId = desc.GetIndexTabletId();
-                        }
-                    }
-                }
-
-                return false;
-            });
-
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        const auto createNodeResponse1 = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"))->Record;
-
-        const auto nodeId1 = createNodeResponse1.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        const auto createNodeResponse2 = service.CreateNode(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file2"))->Record;
-
-        const auto nodeId2 = createNodeResponse2.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((2LU << 56U) + 2, nodeId2);
-
-        bool intercept = true;
-        bool intercepted = false;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvUnlinkNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvUnlinkNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard2Id)
-                    {
-                        intercepted = true;
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        service.SendRenameNodeRequest(
-            headers,
-            RootNodeId,
-            "file1",
-            RootNodeId,
-            "file2",
-            0);
-
-        ui32 iterations = 0;
-        while (!intercepted && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(intercepted);
-        intercept = false;
-
-        auto headers2 = headers;
-        headers2.FileSystemId = shard2Id;
-
-        auto listNodesResponse = service.ListNodes(
-            headers2,
-            shard2Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-
-        TIndexTabletClient tablet(env.GetRuntime(), nodeIdx, tabletId);
-        tablet.RebootTablet();
-
-        auto renameResponse = service.RecvRenameNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_REJECTED,
-            renameResponse->GetError().GetCode(),
-            renameResponse->GetError().GetMessage());
-
-        // remaking session since CreateSessionActor doesn't do it by itself
-        // because EvWakeup never arrives because Scheduling doesn't work by
-        // default and RegistrationObservers get reset after RebootTablet
-        headers = service.InitSession(fsId, "client");
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file2", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        listNodesResponse = service.ListNodes(
-            headers2,
-            shard2Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-    }
-
-    Y_UNIT_TEST(ShouldRetryNodeCreationInShard)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        TAutoPtr<IEventHandle> shardCreateResponse;
-        bool intercept = true;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvCreateNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvCreateNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        auto response = std::make_unique<
-                            TEvService::TEvCreateNodeResponse>(
-                            MakeError(E_REJECTED, "error"));
-
-                        shardCreateResponse = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        const ui64 requestId = 111;
-        service.SendCreateNodeRequest(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"),
-            requestId);
-
-        ui32 iterations = 0;
-        while (!shardCreateResponse && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(shardCreateResponse);
-        intercept = false;
-        env.GetRuntime().Send(shardCreateResponse.Release());
-
-        auto createResponse = service.RecvCreateNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createResponse->GetError().GetCode(),
-            createResponse->GetError().GetMessage());
-
-        const auto nodeId1 = createResponse->Record.GetNode().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        // checking DupCache logic
-        service.SendCreateNodeRequest(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"),
-            requestId);
-
-        createResponse = service.RecvCreateNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createResponse->GetError().GetCode(),
-            createResponse->GetError().GetMessage());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createResponse->Record.GetNode().GetId());
-    }
-
-    Y_UNIT_TEST(ShouldRetryNodeCreationInShardUponLeaderRestart)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        ui64 tabletId = -1;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvDescribeFileStoreResponse: {
-                        using TDesc = TEvSSProxy::TEvDescribeFileStoreResponse;
-                        const auto* msg = event->Get<TDesc>();
-                        const auto& desc =
-                            msg->PathDescription.GetFileStoreDescription();
-                        if (desc.GetConfig().GetFileSystemId() == fsId) {
-                            tabletId = desc.GetIndexTabletId();
-                        }
-                    }
-                }
-
-                return false;
-            });
-
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        bool intercept = true;
-        bool intercepted = false;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvCreateNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvCreateNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        intercepted = true;
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        const ui64 requestId = 111;
-        service.SendCreateNodeRequest(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"),
-            requestId);
-
-        ui32 iterations = 0;
-        while (!intercepted && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(intercepted);
-        intercept = false;
-
-        // TODO listNodes in leader?
-
-        auto headers1 = headers;
-        headers1.FileSystemId = shard1Id;
-
-        auto listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(0, listNodesResponse.NodesSize());
-
-        TIndexTabletClient tablet(env.GetRuntime(), nodeIdx, tabletId);
-        tablet.RebootTablet();
-
-        auto createResponse = service.RecvCreateNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_REJECTED,
-            createResponse->GetError().GetCode(),
-            createResponse->GetError().GetMessage());
-
-        // remaking session since CreateSessionActor doesn't do it by itself
-        // because EvWakeup never arrives because Scheduling doesn't work by
-        // default and RegistrationObservers get reset after RebootTablet
-        // restoreClientSession = true
-        headers = service.InitSession(fsId, "client", {}, true);
-
-        listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        const ui64 nodeId1 = (1LU << 56U) + 2;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        listNodesResponse = service.ListNodes(
-            headers1,
-            shard1Id,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-
-        // checking DupCache logic
-        service.SendCreateNodeRequest(
-            headers,
-            TCreateNodeArgs::File(RootNodeId, "file1"),
-            requestId);
-
-        createResponse = service.RecvCreateNodeResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createResponse->GetError().GetCode(),
-            createResponse->GetError().GetMessage());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createResponse->Record.GetNode().GetId());
-    }
-
-    Y_UNIT_TEST(ShouldRetryNodeCreationInShardUponCreateHandle)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        TAutoPtr<IEventHandle> shardCreateResponse;
-        bool intercept = true;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvCreateNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvCreateNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        auto response = std::make_unique<
-                            TEvService::TEvCreateNodeResponse>(
-                            MakeError(E_REJECTED, "error"));
-
-                        shardCreateResponse = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        const ui64 requestId = 111;
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE,
-            "", // shardId
-            requestId);
-
-        ui32 iterations = 0;
-        while (!shardCreateResponse && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(shardCreateResponse);
-        intercept = false;
-        env.GetRuntime().Send(shardCreateResponse.Release());
-
-        auto createHandleResponse = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createHandleResponse->GetError().GetCode(),
-            createHandleResponse->GetError().GetMessage());
-
-        const auto nodeId1 = createHandleResponse->Record.GetNodeAttr().GetId();
-        UNIT_ASSERT_VALUES_EQUAL((1LU << 56U) + 2, nodeId1);
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        // checking DupCache logic
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE,
-            "", // shardId
-            requestId);
-
-        createHandleResponse = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createHandleResponse->GetError().GetCode(),
-            createHandleResponse->GetError().GetMessage());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse->Record.GetNodeAttr().GetId());
-    }
-
-    Y_UNIT_TEST(ShouldRetryNodeCreationInShardUponCreateHandleUponLeaderRestart)
-    {
-        NProto::TStorageConfig config;
-        config.SetMultiTabletForwardingEnabled(true);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        ui64 tabletId = -1;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvDescribeFileStoreResponse: {
-                        using TDesc = TEvSSProxy::TEvDescribeFileStoreResponse;
-                        const auto* msg = event->Get<TDesc>();
-                        const auto& desc =
-                            msg->PathDescription.GetFileStoreDescription();
-                        if (desc.GetConfig().GetFileSystemId() == fsId) {
-                            tabletId = desc.GetIndexTabletId();
-                        }
-                    }
-                }
-
-                return false;
-            });
-
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ConfigureShards(service, fsId, shard1Id, shard2Id);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        bool intercept = true;
-        bool intercepted = false;
-        env.GetRuntime().SetEventFilter(
-            [&] (auto& runtime, TAutoPtr<IEventHandle>& event) {
-                Y_UNUSED(runtime);
-                if (event->GetTypeRewrite() == TEvService::EvCreateNodeRequest) {
-                    const auto* msg =
-                        event->Get<TEvService::TEvCreateNodeRequest>();
-                    if (intercept
-                            && msg->Record.GetFileSystemId() == shard1Id)
-                    {
-                        intercepted = true;
-                        return true;
-                    }
-                }
-                return false;
-            });
-
-        const ui64 requestId = 111;
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE,
-            "", // shardId
-            requestId);
-
-        ui32 iterations = 0;
-        while (!intercepted && iterations++ < 100) {
-            env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(50));
-        }
-
-        UNIT_ASSERT(intercepted);
-        intercept = false;
-
-        TIndexTabletClient tablet(env.GetRuntime(), nodeIdx, tabletId);
-        tablet.RebootTablet();
-
-        auto createHandleResponse = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            E_REJECTED,
-            createHandleResponse->GetError().GetCode(),
-            createHandleResponse->GetError().GetMessage());
-
-        const auto nodeId1 = (1LU << 56U) + 2;
-
-        // remaking session since CreateSessionActor doesn't do it by itself
-        // because EvWakeup never arrives because Scheduling doesn't work by
-        // default and RegistrationObservers get reset after RebootTablet
-        // restoreClientSession = true
-        headers = service.InitSession(fsId, "client", {}, true);
-
-        auto listNodesResponse = service.ListNodes(
-            headers,
-            fsId,
-            RootNodeId)->Record;
-
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NamesSize());
-        UNIT_ASSERT_VALUES_EQUAL(1, listNodesResponse.NodesSize());
-        UNIT_ASSERT_VALUES_EQUAL("file1", listNodesResponse.GetNames(0));
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            listNodesResponse.GetNodes(0).GetId());
-
-        // checking DupCache logic
-        service.SendCreateHandleRequest(
-            headers,
-            fsId,
-            RootNodeId,
-            "file1",
-            TCreateHandleArgs::CREATE,
-            "", // shardId
-            requestId);
-
-        createHandleResponse = service.RecvCreateHandleResponse();
-        UNIT_ASSERT_VALUES_EQUAL_C(
-            S_OK,
-            createHandleResponse->GetError().GetCode(),
-            createHandleResponse->GetError().GetMessage());
-        UNIT_ASSERT_VALUES_EQUAL(
-            nodeId1,
-            createHandleResponse->Record.GetNodeAttr().GetId());
     }
 
     Y_UNIT_TEST(ShouldUseAliasesForRequestsForwarding)
@@ -6149,924 +3258,6 @@ Y_UNIT_TEST_SUITE(TStorageServiceTest)
     {
         CheckThreeStageWrites(NProto::STORAGE_MEDIA_SSD, true);
         CheckTwoStageReads(NProto::STORAGE_MEDIA_SSD, true);
-    }
-
-    void DoTestShardedFileSystemConfigured(
-        const TString& fsId,
-        TServiceClient& service,
-        TVector<TString> fsList)
-    {
-        auto listing = service.ListFileStores();
-        const auto& fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-
-        UNIT_ASSERT_VALUES_EQUAL(fsList, ids);
-
-        auto headers = service.InitSession(fsId, "client");
-
-        for (const auto& id: ids) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(id);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(1, sessions.size());
-
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.SessionId,
-                sessions[0].GetSessionId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.ClientId,
-                sessions[0].GetClientId());
-            UNIT_ASSERT_VALUES_EQUAL("", sessions[0].GetSessionState());
-        }
-
-        const TString sessionState = "some_state";
-        service.ResetSession(headers, sessionState);
-
-        for (const auto& id: ids) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(id);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(1, sessions.size());
-
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.SessionId,
-                sessions[0].GetSessionId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                headers.ClientId,
-                sessions[0].GetClientId());
-            UNIT_ASSERT_VALUES_EQUAL(
-                sessionState,
-                sessions[0].GetSessionState());
-        }
-
-        service.DestroySession(headers);
-
-        for (const auto& id: ids) {
-            NProtoPrivate::TDescribeSessionsRequest request;
-            request.SetFileSystemId(id);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("describesessions", buf);
-            NProtoPrivate::TDescribeSessionsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-
-            const auto& sessions = response.GetSessions();
-            UNIT_ASSERT_VALUES_EQUAL(0, sessions.size());
-        }
-    }
-
-    Y_UNIT_TEST(ShouldConfigureShardsAutomatically)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 2_GB / 4_KB);
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        DoTestShardedFileSystemConfigured(
-            fsId,
-            service,
-            {fsId, fsId + "_s1", fsId + "_s2"});
-    }
-
-    Y_UNIT_TEST(ShouldNotConfigureShardsAutomaticallyForSmallFileSystems)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, (1_GB - 4_KB) / 4_KB);
-
-        DoTestShardedFileSystemConfigured(fsId, service, {fsId});
-    }
-
-    Y_UNIT_TEST(ShouldHandleErrorsDuringShardedFileSystemCreation)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto blockCount = 2_GB / 4_KB;
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        TVector<TString> expected = {fsId, fsId + "_s1", fsId + "_s2"};
-
-        NProto::TError createShardError;
-        NProto::TError configureShardError;
-        NProto::TError configureShardsError;
-
-        TAutoPtr<IEventHandle> toSend;
-
-        env.GetRuntime().SetEventFilter(
-            [&] (TTestActorRuntimeBase&, TAutoPtr<IEventHandle>& event) {
-                switch (event->GetTypeRewrite()) {
-                    case TEvSSProxy::EvCreateFileStoreRequest: {
-                        using TRequest = TEvSSProxy::TEvCreateFileStoreRequest;
-                        using TResponse =
-                            TEvSSProxy::TEvCreateFileStoreResponse;
-                        const auto* msg = event->Get<TRequest>();
-                        if (msg->Config.GetFileSystemId() != expected[1]) {
-                            break;
-                        }
-
-                        if (!HasError(createShardError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            createShardError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvIndexTablet::EvConfigureAsShardRequest: {
-                        using TRequest =
-                            TEvIndexTablet::TEvConfigureAsShardRequest;
-                        using TResponse =
-                            TEvIndexTablet::TEvConfigureAsShardResponse;
-                        const auto* msg = event->Get<TRequest>();
-                        if (msg->Record.GetFileSystemId() != expected[1]) {
-                            break;
-                        }
-
-                        if (!HasError(configureShardError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            configureShardError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvIndexTablet::EvConfigureShardsRequest: {
-                        using TResponse =
-                            TEvIndexTablet::TEvConfigureShardsResponse;
-
-                        if (!HasError(configureShardsError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            configureShardsError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-
-                return false;
-            });
-
-        createShardError = MakeError(E_REJECTED, "failed to create shard");
-        service.SendCreateFileStoreRequest(fsId, blockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvCreateFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(createShardError),
-                FormatError(response->GetError()));
-        }
-
-        createShardError = {};
-        configureShardError =
-            MakeError(E_REJECTED, "failed to configure shard");
-        service.SendCreateFileStoreRequest(fsId, blockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvCreateFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(configureShardError),
-                FormatError(response->GetError()));
-        }
-
-        configureShardError = {};
-        configureShardsError =
-            MakeError(E_REJECTED, "failed to configure shards");
-        service.SendCreateFileStoreRequest(fsId, blockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvCreateFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(configureShardsError),
-                FormatError(response->GetError()));
-        }
-
-        configureShardsError = {};
-        service.SendCreateFileStoreRequest(fsId, blockCount);
-        {
-            auto response = service.RecvCreateFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(MakeError(S_OK)),
-                FormatError(response->GetError()));
-        }
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        DoTestShardedFileSystemConfigured(
-            fsId,
-            service,
-            {fsId, fsId + "_s1", fsId + "_s2"});
-    }
-
-    Y_UNIT_TEST(ShouldDeleteShardsAutomatically)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId1 = "test1";
-        const TString fsId2 = "test2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId1, 2_GB / 4_KB);
-        service.CreateFileStore(fsId2, 4_GB / 4_KB);
-
-        TVector<TString> expected = {
-            fsId1, fsId1 + "_s1", fsId1 + "_s2",
-            fsId2, fsId2 + "_s1", fsId2 + "_s2", fsId2 + "_s3", fsId2 + "_s4",
-        };
-
-        auto listing = service.ListFileStores();
-        auto fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.DestroyFileStore(fsId1);
-
-        expected = {
-            fsId2, fsId2 + "_s1", fsId2 + "_s2", fsId2 + "_s3", fsId2 + "_s4",
-        };
-
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.DestroyFileStore(fsId2);
-
-        expected = {};
-
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-    }
-
-    Y_UNIT_TEST(ShouldHandleErrorsDuringShardedFileSystemDestruction)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 2_GB / 4_KB);
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        TVector<TString> expected = {fsId, fsId + "_s1", fsId + "_s2"};
-        auto listing = service.ListFileStores();
-        auto fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        NProto::TError getTopologyError;
-        NProto::TError destroyShardError;
-
-        TAutoPtr<IEventHandle> toSend;
-
-        env.GetRuntime().SetEventFilter(
-            [&] (TTestActorRuntimeBase&, TAutoPtr<IEventHandle>& event) {
-                switch (event->GetTypeRewrite()) {
-                    case TEvIndexTablet::EvGetFileSystemTopologyRequest: {
-                        using TResponse =
-                            TEvIndexTablet::TEvGetFileSystemTopologyResponse;
-
-                        if (!HasError(getTopologyError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            getTopologyError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvSSProxy::EvDestroyFileStoreRequest: {
-                        using TRequest = TEvSSProxy::TEvDestroyFileStoreRequest;
-                        using TResponse =
-                            TEvSSProxy::TEvDestroyFileStoreResponse;
-                        const auto* msg = event->Get<TRequest>();
-                        if (msg->FileSystemId != expected[1]) {
-                            break;
-                        }
-
-                        if (!HasError(destroyShardError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            destroyShardError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-
-                return false;
-            });
-
-        getTopologyError = MakeError(E_REJECTED, "failed to get topology");
-        service.SendDestroyFileStoreRequest(fsId);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvDestroyFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(getTopologyError),
-                FormatError(response->GetError()));
-        }
-
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        getTopologyError = {};
-        destroyShardError = MakeError(E_REJECTED, "failed to destroy shard");
-        service.SendDestroyFileStoreRequest(fsId);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvDestroyFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(destroyShardError),
-                FormatError(response->GetError()));
-        }
-
-        expected = {fsId, fsId + "_s1"};
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        destroyShardError = {};
-        service.SendDestroyFileStoreRequest(fsId);
-        {
-            auto response = service.RecvDestroyFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(MakeError(S_OK)),
-                FormatError(response->GetError()));
-        }
-
-        expected = {};
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-    }
-
-    Y_UNIT_TEST(ShouldAddShardsAutomaticallyUponResize)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1_GB / 4_KB);
-
-        TVector<TString> expected = {fsId, fsId + "_s1"};
-        auto listing = service.ListFileStores();
-        auto fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.ResizeFileStore(fsId, 3_GB / 4_KB);
-
-        expected = TVector<TString>{
-            fsId, fsId + "_s1", fsId + "_s2", fsId + "_s3",
-        };
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.ResizeFileStore(fsId, (4_GB - 4_KB) / 4_KB);
-
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.ResizeFileStore(fsId, 4_GB / 4_KB);
-
-        expected = TVector<TString>{
-            fsId, fsId + "_s1", fsId + "_s2", fsId + "_s3", fsId + "_s4"
-        };
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-    }
-
-    Y_UNIT_TEST(ShouldHandleErrorsDuringShardedFileSystemResize)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto newBlockCount = 4_GB / 4_KB;
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-
-        TVector<TString> expected = {fsId, fsId + "_s1", fsId + "_s2"};
-
-        service.CreateFileStore(fsId, 2_GB / 4_KB);
-        auto listing = service.ListFileStores();
-        auto fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        expected = {
-            fsId, fsId + "_s1", fsId + "_s2", fsId + "_s3", fsId + "_s4",
-        };
-
-        NProto::TError getTopologyError;
-        NProto::TError createShardError;
-        NProto::TError configureShardError;
-        NProto::TError configureShardsError;
-
-        TAutoPtr<IEventHandle> toSend;
-
-        env.GetRuntime().SetEventFilter(
-            [&] (TTestActorRuntimeBase&, TAutoPtr<IEventHandle>& event) {
-                switch (event->GetTypeRewrite()) {
-                    case TEvIndexTablet::EvGetFileSystemTopologyRequest: {
-                        using TResponse =
-                            TEvIndexTablet::TEvGetFileSystemTopologyResponse;
-
-                        if (!HasError(getTopologyError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            getTopologyError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvSSProxy::EvCreateFileStoreRequest: {
-                        using TRequest = TEvSSProxy::TEvCreateFileStoreRequest;
-                        using TResponse =
-                            TEvSSProxy::TEvCreateFileStoreResponse;
-                        const auto* msg = event->Get<TRequest>();
-                        if (msg->Config.GetFileSystemId() != expected[3]) {
-                            break;
-                        }
-
-                        if (!HasError(createShardError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            createShardError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvIndexTablet::EvConfigureAsShardRequest: {
-                        using TRequest =
-                            TEvIndexTablet::TEvConfigureAsShardRequest;
-                        using TResponse =
-                            TEvIndexTablet::TEvConfigureAsShardResponse;
-                        const auto* msg = event->Get<TRequest>();
-                        if (msg->Record.GetFileSystemId() != expected[3]) {
-                            break;
-                        }
-
-                        if (!HasError(configureShardError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            configureShardError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-
-                    case TEvIndexTablet::EvConfigureShardsRequest: {
-                        using TResponse =
-                            TEvIndexTablet::TEvConfigureShardsResponse;
-
-                        if (!HasError(configureShardsError)) {
-                            break;
-                        }
-
-                        auto response = std::make_unique<TResponse>(
-                            configureShardsError);
-
-                        toSend = new IEventHandle(
-                            event->Sender,
-                            event->Recipient,
-                            response.release(),
-                            0, // flags
-                            event->Cookie);
-
-                        return true;
-                    }
-                }
-
-                return false;
-            });
-
-        getTopologyError = MakeError(E_REJECTED, "failed to get topology");
-        service.SendResizeFileStoreRequest(fsId, newBlockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvResizeFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(getTopologyError),
-                FormatError(response->GetError()));
-        }
-
-        getTopologyError = {};
-        createShardError = MakeError(E_REJECTED, "failed to create shard");
-        service.SendResizeFileStoreRequest(fsId, newBlockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvResizeFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(createShardError),
-                FormatError(response->GetError()));
-        }
-
-        createShardError = {};
-        configureShardError =
-            MakeError(E_REJECTED, "failed to configure shard");
-        service.SendResizeFileStoreRequest(fsId, newBlockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvResizeFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(configureShardError),
-                FormatError(response->GetError()));
-        }
-
-        configureShardError = {};
-        configureShardsError =
-            MakeError(E_REJECTED, "failed to configure shards");
-        service.SendResizeFileStoreRequest(fsId, newBlockCount);
-        env.GetRuntime().DispatchEvents({}, TDuration::MilliSeconds(100));
-        UNIT_ASSERT(toSend);
-        env.GetRuntime().Send(toSend, nodeIdx);
-        {
-            auto response = service.RecvResizeFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(configureShardsError),
-                FormatError(response->GetError()));
-        }
-
-        configureShardsError = {};
-        service.SendResizeFileStoreRequest(fsId, newBlockCount);
-        {
-            auto response = service.RecvResizeFileStoreResponse();
-            UNIT_ASSERT_VALUES_EQUAL(
-                FormatError(MakeError(S_OK)),
-                FormatError(response->GetError()));
-        }
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        DoTestShardedFileSystemConfigured(fsId, service, expected);
-    }
-
-    Y_UNIT_TEST(ShouldValidateShardList)
-    {
-        NProto::TStorageConfig config;
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const auto shard1Id = fsId + "-f1";
-        const auto shard2Id = fsId + "-f2";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1'000);
-        service.CreateFileStore(shard1Id, 1'000);
-        service.CreateFileStore(shard2Id, 1'000);
-
-        ui32 shardNo = 0;
-        for (const auto& shardId: {shard1Id, shard2Id}) {
-            NProtoPrivate::TConfigureAsShardRequest request;
-            request.SetFileSystemId(shardId);
-            request.SetShardNo(++shardNo);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureasshard", buf);
-            NProtoPrivate::TConfigureAsShardResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard1Id;
-            *request.AddShardFileSystemIds() = shard2Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            auto jsonResponse = service.ExecuteAction("configureshards", buf);
-            NProtoPrivate::TConfigureShardsResponse response;
-            UNIT_ASSERT(google::protobuf::util::JsonStringToMessage(
-                jsonResponse->Record.GetOutput(), &response).ok());
-        }
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        {
-            auto response = service.CreateSession(THeaders{fsId, "client", ""});
-            const auto& fileStore = response->Record.GetFileStore();
-            UNIT_ASSERT_VALUES_EQUAL(2, fileStore.ShardFileSystemIdsSize());
-            UNIT_ASSERT_VALUES_EQUAL(
-                shard1Id,
-                fileStore.GetShardFileSystemIds(0));
-            UNIT_ASSERT_VALUES_EQUAL(
-                shard2Id,
-                fileStore.GetShardFileSystemIds(1));
-        }
-
-        // shard order change not allowed
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard2Id;
-            *request.AddShardFileSystemIds() = shard1Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_ARGUMENT,
-                response->GetStatus(),
-                response->GetErrorReason());
-        }
-
-        // shard list cropping not allowed
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            *request.AddShardFileSystemIds() = shard1Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_ARGUMENT,
-                response->GetStatus(),
-                response->GetErrorReason());
-        }
-
-        // force flag should override checks
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(fsId);
-            request.SetForce(true);
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                S_OK,
-                response->GetStatus(),
-                response->GetErrorReason());
-        }
-
-        // waiting for IndexTablet start after the restart triggered by
-        // configureshards
-        WaitForTabletStart(service);
-
-        {
-            auto response = service.CreateSession(THeaders{fsId, "client", ""});
-            const auto& fileStore = response->Record.GetFileStore();
-            UNIT_ASSERT_VALUES_EQUAL(0, fileStore.ShardFileSystemIdsSize());
-        }
-
-        // configureshards for a shard should fail
-        {
-            NProtoPrivate::TConfigureShardsRequest request;
-            request.SetFileSystemId(shard1Id);
-            *request.AddShardFileSystemIds() = shard2Id;
-
-            TString buf;
-            google::protobuf::util::MessageToJsonString(request, &buf);
-            service.SendExecuteActionRequest("configureshards", buf);
-            auto response = service.RecvExecuteActionResponse();
-            UNIT_ASSERT_VALUES_EQUAL_C(
-                E_INVALID_STATE,
-                response->GetStatus(),
-                response->GetErrorReason());
-        }
-    }
-
-    Y_UNIT_TEST(ShouldNotAutomaticallyAddShardsToShards)
-    {
-        NProto::TStorageConfig config;
-        config.SetAutomaticShardCreationEnabled(true);
-        config.SetShardAllocationUnit(1_GB);
-        config.SetAutomaticallyCreatedShardSize(2_GB);
-        TTestEnv env({}, config);
-        env.CreateSubDomain("nfs");
-
-        ui32 nodeIdx = env.CreateNode("nfs");
-
-        const TString fsId = "test";
-        const TString shard1Id = fsId + "_s1";
-
-        TServiceClient service(env.GetRuntime(), nodeIdx);
-        service.CreateFileStore(fsId, 1_GB / 4_KB);
-
-        TVector<TString> expected = {fsId, shard1Id};
-        auto listing = service.ListFileStores();
-        auto fsIds = listing->Record.GetFileStores();
-        TVector<TString> ids(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
-
-        service.ResizeFileStore(shard1Id, 4_GB / 4_KB);
-
-        // subshards shouldn't have been created for shard1
-        expected = TVector<TString>{fsId, shard1Id};
-        listing = service.ListFileStores();
-        fsIds = listing->Record.GetFileStores();
-        ids = TVector<TString>(fsIds.begin(), fsIds.end());
-        Sort(ids);
-        UNIT_ASSERT_VALUES_EQUAL(expected, ids);
     }
 }
